+++ resources/views/admin/user/index.blade.php
@@ -11,24 +11,28 @@
                 @endif
             </div>
             <!-- /.col-lg-12 -->
-             <a href="{{url('admin/users/create')}}" class="btn btn-primary">{{trans('text.add_user')}}</a>
-            <table class="table table-striped table-bordered table-hover" id="dataTables-example">
+            <table class="table table-striped table-bordered table-hover" >
                 <thead>
-                    <tr align="center">
-                        <th>{{trans('text.choose')}}</th>
-                        <th>{{trans('text.username')}}</th>
-                        <th>{{trans('text.email')}}</th>
-                        <th>{{trans('text.division')}}</th>
-                        <th>{{trans('text.delete')}}</th>
-                        <th>{{trans('text.edit')}}</th>
+                    <tr align="center" >
+                        <th class="center">{{trans('text.choose_to_reset_pass')}}</th>
+                        <th class="center">{{trans('text.username')}}</th>
+                        <th class="center">{{trans('text.email')}}</th>
+                        <th class="center">{{trans('text.active')}}</th>
+                        <th class="center">{{trans('text.division')}}</th>
+                        <th class="center">{{trans('text.delete')}}</th>
+                        <th class="center">{{trans('text.edit')}}</th>
                     </tr>
                 </thead>
                 <tbody>
                     @foreach($users as $user)
                         <tr class="odd gradeX" align="center" id = "user_{{$user->id}}">
                             <td><input type="checkbox" value="{{$user->id}}" class="choose-to-resetPassword"></td>
-                            <td>{{$user->name}}</td>
+                            <td><a href="{{url('admin/users/'.$user->id)}}">{{$user->name}}</a></td>
                             <td>{{$user->email}}</td>
+                            <td>
+                                @if($user->isActive)
+                                    <i class="fa fa-check" aria-hidden="true"></i></td>
+                                @endif
                             <td>
                                 @if($user->belongDivision)
                                     {{$user->belongDivision->name}}
@@ -51,6 +55,7 @@
             </table>
             <button class="btn btn-primary click-to-reset">{{trans('text.reset_password')}}</button>
             <a href="{{route('exportExcel')}}" class="btn btn-primary">{{trans('text.export_to_excel')}}</a>
+            <a href="{{url('admin/users/create')}}" class="btn btn-primary">{{trans('text.add_user')}}</a>
         </div>
         <!-- /.row -->
     </div>
@@ -62,19 +67,22 @@
     <script>
         $(document).ready(function(){
             $('.delete-user-btn').click(function(){
-                $.ajaxSetup({
+                var check = confirm("Want to delete?");
+                if (check) {
+                    $.ajaxSetup({
                     headers: {
                     'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
                     }
-                });
-                var data = $(this).attr('data');
-                $.ajax({
-                    url: '/admin/users/'+data,
-                    type: 'DELETE',  // user.destroy
-                    success: function(result) {  
-                    }
-                });
-                $(this).closest('tr').hide();
+                    });
+                    var data = $(this).attr('data');
+                    $.ajax({
+                        url: '/admin/users/'+data,
+                        type: 'DELETE',  // user.destroy
+                        success: function(result) {  
+                        }
+                    });
+                    $(this).closest('tr').hide();
+                }
             });
 
             $('.click-to-reset').click(function(){
@@ -92,6 +100,8 @@
                   data: list_id,
                 },
                 function(respont){
+                    alert('ok');
+                    location.reload();
                 });
             });
         });
